Added ChangeHTML function

+++ BlazorProcesses/NonClassFunctions.py
@@ -37,7 +37,7 @@ def PublishComponents(directory: str) -> bool:
     else:
         return False
 
-def CopyAndPasteFolders(source: str, destination: str):
+def CopyAndPasteFolders(source: str, destination: str) -> bool:
     folders_to_copy = ['_content','_framework']
     success = False
     try:
@@ -45,15 +45,47 @@ def CopyAndPasteFolders(source: str, destination: str):
             source_folder = os.path.join(source, folder)
             destination_folder = os.path.join(destination, folder)
             if os.path.isdir(source_folder):
-                shutil.copytree(source_folder, destination_folder)
-                print(f"Successfully copied folder '{folder}' from '{source}' to '{destination}'.")
+                if os.path.exists(destination_folder):
+                    print(f"Folder '{folder}' already exists in '{destination}', skipping copy.")
+                else:
+                    shutil.copytree(source_folder, destination_folder)
+                    print(f"Successfully copied folder '{folder}' from '{source}' to '{destination}'.")
                 success = True
-                return success
             else:
                 print(f"Folder '{folder}' not found in '{source}'.")
-                return success
+        return success
     except Exception as e:
         print(f"Error copying folder(s) from '{source}' to '{destination}': {e}")
         return success
 
+def ChangeHTML(path: str):
+    variable_name = "</body>"
+    variable_line = None
+    line1 = f'<script src="_framework/blazor.webassembly.js"></script>'
+    line2 = f'<script src="_content/Microsoft.AspNetCore.Components.CustomElements/Microsoft.AspNetCore.Components.CustomElements.lib.module.js"></script>'
+
+    try:
+        with open(path, "r") as file:
+            contents = file.read()
+            lines = contents.splitlines()
+        
+        for i, line in enumerate(lines):
+            if line1 in line or line2 in line:
+                print("Scripts are already in the HTML file")
+                return True
+            if variable_name in line:
+                variable_line = i
+                break
+
+        with open(path, "r+") as file:
+            file.seek(0)
+            file.write('\n'.join(lines[:variable_line]) + '\n' + line1 + '\n' + line2 + '\n'.join(lines[variable_line:]))
+            file.truncate()
+        
+        return True
+
+
+    except Exception as e:
+        print(f"Error: {e}")
+        return False
